A3: move blackjack state/persistence into services/, break actions<->_state cycle
