+++ disbot/views/blackjack/pvp_view.py
@@ -10,8 +10,11 @@ from __future__ import annotations
 
 import discord
 
-from cogs.blackjack._persistence import _clear_pvp_match, _save_pvp_match
-from cogs.blackjack._state import (
+from core.runtime.interaction_helpers import safe_edit
+from services import economy_service, game_wager_workflow
+from services.blackjack_engine import is_blackjack as _is_blackjack
+from services.blackjack_persistence import _clear_pvp_match, _save_pvp_match
+from services.blackjack_state import (
     BLACKJACK_PVP_ESCROW_SUBSYSTEM,
     BLACKJACK_PVP_ESCROW_VERSION,
     _active,
@@ -19,9 +22,6 @@ from cogs.blackjack._state import (
     _pvp,
     _PvPState,
 )
-from core.runtime.interaction_helpers import safe_edit
-from services import economy_service, game_wager_workflow
-from services.blackjack_engine import is_blackjack as _is_blackjack
 from utils.ui_constants import ECONOMY_COLOR, GAME_COLOR
 from views.base import handle_view_error as _on_view_error
 from views.blackjack.embeds import _game_embed
