+++ disbot/views/blackjack/solo_view.py
@@ -23,12 +23,12 @@ from __future__ import annotations
 
 import discord
 
-from cogs.blackjack._persistence import _clear_solo_game, _save_game_state
-from cogs.blackjack._state import FREE_WIN_COINS, _active, _Game
 from core.runtime.interaction_helpers import safe_defer, safe_edit, safe_followup
 from services import economy_service
 from services.blackjack_engine import hand_value as _hand_value
 from services.blackjack_engine import is_blackjack as _is_blackjack
+from services.blackjack_persistence import _clear_solo_game, _save_game_state
+from services.blackjack_state import FREE_WIN_COINS, _active, _Game
 from utils import db
 from utils.ui_constants import ECONOMY_COLOR, ERROR_COLOR, GAME_COLOR, SUCCESS_COLOR
 from views.base import handle_view_error as _on_view_error
