+++ disbot/views/blackjack/tournament_views.py
@@ -18,7 +18,11 @@ import logging
 import discord
 from discord.ext import commands
 
-from cogs.blackjack._state import (
+from core.runtime import resources
+from services import game_wager_workflow, tournament_state_service
+from services.blackjack_engine import hand_value as _hand_value
+from services.blackjack_engine import is_blackjack as _is_blackjack
+from services.blackjack_state import (
     BLACKJACK_TOURNAMENT_SUBSYSTEM,
     TOURN_BET_PER_ROUND,
     _active,
@@ -27,10 +31,6 @@ from cogs.blackjack._state import (
     _tournaments,
     _TournPlayerState,
 )
-from core.runtime import resources
-from services import game_wager_workflow, tournament_state_service
-from services.blackjack_engine import hand_value as _hand_value
-from services.blackjack_engine import is_blackjack as _is_blackjack
 from utils.channels import cleanup_category
 from utils.ui_constants import ECONOMY_COLOR, ERROR_COLOR, GAME_COLOR, SUCCESS_COLOR
 from views.base import handle_view_error as _on_view_error
